CHORE: Mudança das rotas dos components kitchen
