--- src/app/components/kitchen/kitchen-ready/kitchen-ready.module.ts
@@ -1,15 +1,0 @@
-import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
-import { KitchenReadyComponent } from './kitchen-ready.component';
-
-
-
-@NgModule({
-  declarations: [
-    KitchenReadyComponent
-  ],
-  imports: [
-    CommonModule
-  ]
-})
-export class KitchenReadyModule { }
